azure-ai/functions/requirements.txt
azure-functions==1.21.3
azure-search-documents==11.6.0
openai==1.75.0
